Initial Kue work
This commit moves planet-stream to an entirely different queuing system
managed by "Kue" (https://github.com/Automattic/kue). This allows for better
logging of failed and completed syncs of changesets between overpass and
planet.osm.org.

Furthermore, kue brings a UI to the tasks to see the state of the processing
pipeline with kue_dashboard. If running inside a docker container, you can
use the command `node_modules/kue/bin/kue-dashboard -r redis://DOCKER_HOST:REDIS_PORT`
to access the UI.

A new file `kue_cleanup` will run at an interval of 5 minutes and cleanup tasks
that are completed, failed or active if they are older than a certain length
of time. For completed and failed tasks it's specified to be 30 days, and for active
it's 1 day.

I also bumped the node version for the dockerfile to 6.0